tests/wpt/web-platform-tests/css/css-box/parsing/overflow-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS basic box model: parsing overflow with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-box-3/#propdef-overflow">
<meta name="assert" content="overflow supports only the grammar '[ visible | hidden | scroll | auto | no-display | no-content ]{1,2}'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("overflow", "none");
test_invalid_value("overflow", "visible hidden scroll");
test_invalid_value("overflow", "no-display, no-content");

test_invalid_value("overflow-x", "none");
test_invalid_value("overflow-y", "visible hidden");
test_invalid_value("overflow-y", "no-display, no-content");
</script>
</body>
</html>
